Fix @azureBaseType decorator: move to BaseTypes namespace, use rest params
- Move extern dec from decorators.tsp to base-types/base-types.tsp to avoid
  breaking common-types compilation (which imports decorators.tsp standalone)
- Change signature from valueof BaseTypeInfo[] to ...valueof BaseTypeInfo[]
  for cleaner TS types and simpler call syntax
- Add Azure.ResourceManager.BaseTypes namespace exclusion to
  genericResourceInternal decorator check
- Remove circular imports from base-types/*.tsp files
- Update tests to use new syntax and add using statement
- Regenerate docs, samples, and extern signatures
