+++ packages/samples/specs/resource-manager/resource-types/agent/main.tsp
@@ -26,6 +26,7 @@ enum Versions {
 
 /** RP-specific properties for the Contoso agent (Platform deployment model) */
 model ContosoAgentProperties is AgentPropertiesPlatform {
+  ...ModelDeploymentRefProperty;
   ...DefaultProvisioningStateProperty;
 }
