+++ packages/typespec-azure-resource-manager/generated-defs/Azure.ResourceManager.BaseTypes.ts-test.ts
@@ -1,0 +1,11 @@
+// An error in the imports would mean that the decorator is not exported or
+// doesn't have the right name.
+
+import { $decorators } from "@azure-tools/typespec-azure-resource-manager";
+import type { AzureResourceManagerBaseTypesDecorators } from "./Azure.ResourceManager.BaseTypes.js";
+
+/**
+ * An error here would mean that the exported decorator is not using the same signature. Make sure to have export const $decName: DecNameDecorator = (...) => ...
+ */
+const _decs: AzureResourceManagerBaseTypesDecorators =
+  $decorators["Azure.ResourceManager.BaseTypes"];
