+++ website/src/content/docs/docs/libraries/azure-resource-manager/reference/index.mdx
@@ -46,7 +46,6 @@ npm install --save-peer @azure-tools/typespec-azure-resource-manager
 - [`@armResourceRead`](./decorators.md#@Azure.ResourceManager.armResourceRead)
 - [`@armResourceUpdate`](./decorators.md#@Azure.ResourceManager.armResourceUpdate)
 - [`@armVirtualResource`](./decorators.md#@Azure.ResourceManager.armVirtualResource)
-- [`@azureBaseType`](./decorators.md#@Azure.ResourceManager.azureBaseType)
 - [`@extensionResource`](./decorators.md#@Azure.ResourceManager.extensionResource)
 - [`@identifiers`](./decorators.md#@Azure.ResourceManager.identifiers)
 - [`@locationResource`](./decorators.md#@Azure.ResourceManager.locationResource)
@@ -183,6 +182,10 @@ npm install --save-peer @azure-tools/typespec-azure-resource-manager
 
 ## Azure.ResourceManager.BaseTypes
 
+### Decorators
+
+- [`@azureBaseType`](./decorators.md#@Azure.ResourceManager.BaseTypes.azureBaseType)
+
 ### Models
 
 - [`BaseTypeInfo`](./data-types.md#Azure.ResourceManager.BaseTypes.BaseTypeInfo)
@@ -203,8 +206,6 @@ npm install --save-peer @azure-tools/typespec-azure-resource-manager
 - [`AgentResponse`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.AgentResponse)
 - [`AgentToolProperty`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.AgentToolProperty)
 - [`AgentToolTypeEntry`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.AgentToolTypeEntry)
-- [`AgentToolTypeEntryAppliance`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.AgentToolTypeEntryAppliance)
-- [`AgentToolTypeEntryPlatform`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.AgentToolTypeEntryPlatform)
 - [`ConversationOutput`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.ConversationOutput)
 - [`ConversationProperties`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.ConversationProperties)
 - [`InputMessage`](./data-types.md#Azure.ResourceManager.BaseTypes.Agents.InputMessage)
